PLATOPS-1329: Move actions to dedicated package
