+++ app/uk/gov/hmrc/cataloguefrontend/actions/VerifySignInStatus.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.cataloguefrontend
+package uk.gov.hmrc.cataloguefrontend.actions
 
 import javax.inject.Inject
 import play.api.mvc._
