+++ app/views/digital_service_info.scala.html
@@ -20,7 +20,7 @@
 @import uk.gov.hmrc.cataloguefrontend.UserManagementConnector.TeamMember
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages._
 @import uk.gov.hmrc.cataloguefrontend.DisplayableTeamMember
-@import uk.gov.hmrc.cataloguefrontend.UmpAuthRequest
+@import uk.gov.hmrc.cataloguefrontend.actions.UmpAuthRequest
 
 @(timestampedRepos: DigitalServiceDetails,
   digitalServiceOwner: Option[DisplayableTeamMember]
